Speed up eForge::process_bits
There is a very small gain in using substr rather than split. The major gain is in
the new "if" clause within the foreach loop. It is only true if the bit is 1 (as the
bit can only be 0 or 1) and will therefore increment the counter for that cell and
push this probe to the list of MVPs for that cell.