FDR correction + similar dPlot and PDF
The points are now colored according to the q-value instead. Also, the PDF output has changed
to be more similar to the dimple plot. The dimple plot now is also more similar to the PDF
as the cell lines/types are sorted in the same way and the break lines between tissues are
also shown.